timeleft refactor
