+++ src/App.js
@@ -8,7 +8,7 @@ import {
 import "./App.css";
 import { auth, app } from "./firebase/firebase_config";
 import useFirestore from "./hooks/useFirestore";
-import useForceUpdate from "./hooks/useForceUpdate";
+
 import firebase from "firebase/compat/app";
 import loginIcon from "./assets/login 1.png";
 import userIcon from "./assets/user 1.svg";
@@ -17,6 +17,7 @@ import moment from "moment";
 import Button from "./UI/Button";
 import Login from "./routes/Login";
 import Register from "./routes/Register";
+import TimeLeft from "./components/TimeLeft";
 
 import { BrowserRouter as Router, Routes, Route, Link } from "react-router-dom";
 
@@ -188,77 +189,5 @@ function App() {
     </Router>
   );
 }
-
-const TimeLeft = ({ shoutOutTime }) => {
-  const forceUpdate = useForceUpdate();
-  useEffect(() => {
-    forceUpdate();
-  }, [shoutOutTime]);
-
-  const calculateTimeLeft = () => {
-    const difference =
-      new Date(
-        moment.utc(shoutOutTime).local().format("YYYY-MM-DD HH:mm:ss")
-      ).getTime() - new Date().getTime();
-    let timeLeft = {};
-    if (difference > 0) {
-      timeLeft = {
-        days: Math.floor(difference / (1000 * 60 * 60 * 24)),
-        hours: Math.floor((difference / (1000 * 60 * 60)) % 24),
-        minutes: Math.floor((difference / 1000 / 60) % 60),
-        seconds: Math.floor((difference / 1000) % 60),
-      };
-    }
-
-    return timeLeft;
-  };
-
-  const [timeLeft, setTimeLeft] = useState(calculateTimeLeft());
-  const [year] = useState(new Date().getFullYear());
-
-  useEffect(() => {
-    setTimeout(() => {
-      setTimeLeft(calculateTimeLeft());
-    }, 1000);
-  });
-
-  const timerComponents = [];
-
-  Object.keys(timeLeft).forEach((interval, index) => {
-    if (!timeLeft[interval]) {
-      return;
-    }
-
-    timerComponents.push(
-      <span key={index}>
-        <span className="interval-number">
-          {timeLeft[interval].toString().length > 1 ? (
-            <>
-              <span className="timer-number">
-                {timeLeft[interval].toString().split("")[0]}
-              </span>
-              <span className="timer-number">
-                {timeLeft[interval].toString().split("")[1]}
-              </span>
-            </>
-          ) : (
-            <>
-              <span className="timer-number">0</span>
-              <span className="timer-number">{timeLeft[interval]}</span>
-            </>
-          )}
-        </span>
-        <span className="interval-sign">{interval}</span>
-      </span>
-    );
-  });
-  return (
-    <>
-      <span className="timer-wrapper">
-        {timerComponents.length > 0 ? timerComponents : "0 min"}{" "}
-      </span>
-    </>
-  );
-};
 
 export default App;
